add: round trip numeric tests fix: write decimal as decimal, not double fix: add ulong support, for unsigned numbers > long.MaxValue fix: throw exception on failed numeric parsing, safer than silently returning 0 fix: use "R" numeric format when writing double, for round-trip stability
TODO: still failing RoundTripFloatDecimal test, will check to see if it works in other branch

+++ Source/JsonMapper.cs
@@ -534,13 +534,13 @@ public class JsonMapper {
 			writer.Write(Convert.ToInt64((uint)obj));
 		};
 		baseExportTable[typeof(ulong)] = delegate(object obj, JsonWriter writer) {
-			writer.Write(Convert.ToInt64((ulong)obj));
+			writer.Write(Convert.ToUInt64((ulong)obj));
 		};
 		baseExportTable[typeof(float)] = delegate(object obj, JsonWriter writer) {
 			writer.Write(Convert.ToDouble((float)obj));
 		};
 		baseExportTable[typeof(decimal)] = delegate(object obj, JsonWriter writer) {
-			writer.Write(Convert.ToDouble((decimal)obj));
+			writer.Write(Convert.ToDecimal((decimal)obj));
 		};
 		baseExportTable[typeof(DateTime)] = delegate(object obj, JsonWriter writer) {
 			writer.Write(Convert.ToString((DateTime)obj, datetimeFormat));
